+++ crates/bytecode/src/compiler.rs
@@ -3126,15 +3126,27 @@ impl Compiler {
             self.push_register()?
         };
 
-        for arg in args.iter() {
+        let mut packed_arg_indices = AstVec::<u8>::new();
+        for (i, arg) in args.iter().enumerate() {
+            let arg = if let Node::PackedExpression(packed_arg) = ctx.node(*arg) {
+                packed_arg_indices.push(i as u8);
+                packed_arg
+            } else {
+                arg
+            };
             let arg_register = self.push_register()?;
             self.compile_node(*arg, ctx.with_fixed_register(arg_register))?;
         }
-
         if let Some(piped_arg) = piped_arg {
             arg_count += 1;
             let arg_register = self.push_register()?;
             self.push_op(Copy, &[arg_register, piped_arg]);
+        }
+
+        // Indices of args that need to be unpacked are placed in the registers following the args
+        for index in packed_arg_indices.iter() {
+            let register = self.push_register()?;
+            self.push_op(SetNumberU8, &[register, *index]);
         }
 
         let call_result_register = if let Some(result_register) = result.register {
@@ -3154,6 +3166,7 @@ impl Compiler {
                     instance_register,
                     frame_base,
                     arg_count as u8,
+                    packed_arg_indices.len() as u8,
                 ],
             );
         } else {
@@ -3164,6 +3177,7 @@ impl Compiler {
                     function_register,
                     frame_base,
                     arg_count as u8,
+                    packed_arg_indices.len() as u8,
                 ],
             );
         }