+++ crates/runtime/tests/vm_tests.rs
@@ -1328,46 +1328,119 @@ f [10, (1, 2, 3), 20, 30]
             }
         }
 
-        #[test]
-        fn variadic_function() {
-            let script = "
+        mod variadic {
+            use super::*;
+
+            #[test]
+            fn with_two_extra_args() {
+                let script = "
 f = |a, b, c...|
   a + b + c.fold 0, |x, y| x + y
 f 5, 10, 20, 30";
-            check_script_output(script, 65);
-        }
+                check_script_output(script, 65);
+            }
 
-        #[test]
-        fn variadic_function_stacked_call() {
-            let script = "
+            #[test]
+            fn with_three_extra_args() {
+                let script = "
+f = |a, b...| b
+f 42, 1, 2, 3";
+                check_script_output(script, number_tuple(&[1, 2, 3]));
+            }
+
+            #[test]
+            fn with_no_extra_args() {
+                let script = "
+f = |a, b...| b
+f 42";
+                check_script_output(script, tuple(&[]));
+            }
+
+            #[test]
+            fn stacked_call() {
+                let script = "
 f = |a, b, c...|
   a + b + c.fold 0, |x, y| x + y
 f (f 5, 10, 20, 30), 40, 50";
-            check_script_output(script, 155);
-        }
+                check_script_output(script, 155);
+            }
 
-        #[test]
-        fn variadic_function_without_contents() {
-            let script = "
-f = |a, b...| b
-f 42";
-            check_script_output(script, tuple(&[]));
-        }
-
-        #[test]
-        fn variadic_function_with_contents() {
-            let script = "
-f = |a, b...| b
-f 42, 1, 2, 3";
-            check_script_output(script, number_tuple(&[1, 2, 3]));
-        }
-
-        #[test]
-        fn variadic_function_after_default_arg() {
-            let script = "
+            #[test]
+            fn variadic_arg_after_default_arg() {
+                let script = "
 f = |a = 3, b...| b
 f()";
-            check_script_output(script, tuple(&[]));
+                check_script_output(script, tuple(&[]));
+            }
+        }
+
+        mod packed_call_args {
+            use super::*;
+
+            #[test]
+            fn variadic_forwarding() {
+                let script = "
+f = |args...| args
+g = |args...| f args...
+g 1, 2, 3
+";
+                check_script_output(script, number_tuple(&[1, 2, 3]));
+            }
+
+            #[test]
+            fn two_unpacked_args() {
+                let script = "
+f = |a, b, c, d| (a, b, c, d)
+x = (1, 2)
+y = (3, 4)
+f x..., y...
+";
+                check_script_output(script, number_tuple(&[1, 2, 3, 4]));
+            }
+
+            #[test]
+            fn unpacking_before_and_after_normal_arg() {
+                let script = "
+f = |args...| args.to_string()
+x = 'ab'
+y = 'de'
+f x..., 'c', y...
+";
+                check_script_output(script, "abcde".to_string());
+            }
+
+            #[test]
+            fn with_piped_call() {
+                let script = "
+f = |args...| args
+x = (1, 2)
+y = (3, 4)
+5 -> f x..., y...
+";
+                check_script_output(script, number_tuple(&[1, 2, 3, 4, 5]));
+            }
+
+            #[test]
+            fn with_stateful_iterators() {
+                let script = "
+f = |a, b, c, d| (a, b, c, d)
+m = {n: 0}
+g = ||
+  loop
+    yield m.n += 1
+f g().take(2)..., g().take(2)...
+";
+                check_script_output(script, number_tuple(&[1, 2, 3, 4]));
+            }
+
+            #[test]
+            fn with_empty_packed_args() {
+                let script = "
+f = |a, b, c, d| (a, b, c, d)
+f []..., (1, 2, 3, 4)...
+";
+                check_script_output(script, number_tuple(&[1, 2, 3, 4]));
+            }
         }
 
         #[test]